code refactoring for edit and serve order journey
